// admin-ui/src/graphql/queries/types/GetDomains.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: GetDomains
// ====================================================

export interface GetDomains_settings {
  __typename: 'Settings';
  authAllowedDomains: string[];
}

export interface GetDomains {
  settings: GetDomains_settings;
}